Move moderation constants into their own file so they can be imported without importing models
