+++ moderation/constants.py
@@ -1,0 +1,6 @@
+MODERATION_READY_STATE = 0
+MODERATION_DRAFT_STATE = 1
+
+MODERATION_STATUS_REJECTED = 0
+MODERATION_STATUS_APPROVED = 1
+MODERATION_STATUS_PENDING = 2
